requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.10.2
cssselect2==0.8.0
ddddocr==1.0.6
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.5.3
flatbuffers==25.2.10
fonttools==4.58.5
gevent==25.5.1
git-filter-repo==2.47.0
greenlet==3.2.3
humanfriendly==10.0
icalendar==6.3.1
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
lxml==6.0.0
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydyf==0.11.0
Pygments==2.19.2
PyMuPDF==1.26.3
PyPDF2==3.0.1
pyphen==0.17.2
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.4
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zope.event==5.1
zope.interface==7.2
zopfli==0.2.3.post1

# OAuth认证相关包
django-allauth==65.10.0
requests-oauthlib==2.0.0
PyJWT==2.10.1
cryptography==45.0.6

# 可选（提高 DOCX 页数精度）：
# pywin32==306         # Windows + 已安装 Word 时启用 COM 统计页数
# docx2pdf==0.1.8      # 将 DOCX 转成 PDF 后读取页数（通常同样需要 Word）
